requirements.txt
requests
beautifulsoup4
streamlit
googlesearch-python
pandas